Create some helpers for cleaning the code
